// src/lib/types/generated/SshResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TunnelInfo } from "./TunnelInfo";

/**
 * An SSH call's result. `close` gives `null`.
 */
export type SshResponse = { "method": "open", "result": TunnelInfo } | { "method": "close", "result": null };